+++ src/app/body/create-new-organization/create-new-organization.component.ts
@@ -119,25 +119,15 @@ export class CreateNewOrganizationComponent implements OnInit {
   }
 
   checkOrgDomainAvailabilityLive() {
-    const orgDomain = this.orgDomain.value;
-    const callable = this.functions.httpsCallable('teams/orgDomainCheck');
-       callable({OrganizationDomain: orgDomain}).subscribe({
-        next: (result) => {
-          if(result == "orgDomain Already taken"){   
-            this.orgDomainsAvailable = false;
-
-          } else {
-            this.orgDomainsAvailable = true;
-            this.enableLoader = false;
-          }
-        },
-        error: (error) => {
-          console.log(error);
-          console.log("error is here")
-          this.errorHandlerService.showError = true;
-          this.errorHandlerService.getErrorCode(this.componentName, "InternalError","Api");
-        },
-        complete: () => {console.info('Successful ')}
-    });
+    for (let i = 0; i < this.authService.allOrgDomains.length; i++) {
+      if (this.authService.allOrgDomains[i]==this.orgDomain.value) {
+        this.orgDomainsAvailable = false;
+        break;
+      }else {
+        this.orgDomainsAvailable = true;
+          this.enableLoader = false;
+      }
+  }
+  
   }
 }
